Move Card markup to a separate component
