+++ modules/ppcp-axo-block/resources/js/components/Payment.js
@@ -1,14 +1,5 @@
-import { useEffect, useCallback, useMemo } from '@wordpress/element';
-
-const cardIcons = {
-	VISA: 'visa-light.svg',
-	MASTER_CARD: 'mastercard-light.svg',
-	AMEX: 'amex-light.svg',
-	DISCOVER: 'discover-light.svg',
-	DINERS: 'dinersclub-light.svg',
-	JCB: 'jcb-light.svg',
-	UNIONPAY: 'unionpay-light.svg',
-};
+import { useEffect, useCallback } from '@wordpress/element';
+import { CreditCard } from './CreditCard';
 
 export const Payment = ( {
 	fastlaneSdk,
@@ -17,9 +8,6 @@ export const Payment = ( {
 	isGuest,
 	onPaymentLoad,
 } ) => {
-	const { brand, lastDigits, expiry } = card?.paymentSource?.card ?? {};
-	const { fullName } = shippingAddress?.name ?? {};
-
 	// Memoized Fastlane card rendering
 	const loadPaymentComponent = useCallback( async () => {
 		if ( isGuest ) {
@@ -35,40 +23,14 @@ export const Payment = ( {
 		loadPaymentComponent();
 	}, [ loadPaymentComponent ] );
 
-	// Memoized card logo rendering
-	const cardLogo = useMemo( () => {
-		return cardIcons[ brand ] ? (
-			<img
-				className="wc-block-axo-block-card__meta-icon"
-				title={ brand }
-				src={ `${ window.wc_ppcp_axo.icons_directory }${ cardIcons[ brand ] }` }
-				alt={ brand }
-			/>
-		) : (
-			<span>{ brand }</span>
-		);
-	}, [ brand ] );
-
-	const formattedExpiry = expiry
-		? `${ expiry.split( '-' )[ 1 ] }/${ expiry.split( '-' )[ 0 ] }`
-		: '';
-
 	return isGuest ? (
 		<div id="fastlane-card" key="fastlane-card" />
 	) : (
-		<div key="custom-card" className="wc-block-checkout-axo-block-card">
-			<div className="wc-block-checkout-axo-block-card__meta-container">
-				<div className="wc-block-axo-block-card__meta">
-					<span className="wc-block-axo-block-card__meta__digits">
-						{ `**** **** **** ${ lastDigits }` }
-					</span>
-					{ cardLogo }
-				</div>
-				<div className="wc-block-axo-block-card__meta">
-					<span>{ fullName }</span>
-					<span>{ formattedExpiry }</span>{ ' ' }
-				</div>
-			</div>
-		</div>
+		<CreditCard
+			key="custom-card"
+			card={ card }
+			shippingAddress={ shippingAddress }
+			fastlaneSdk={ fastlaneSdk }
+		/>
 	);
 };
